chore(5387): rename table
